+++ src/conic_form.jl
@@ -30,7 +30,7 @@ function +(c::ConicObj, d::ConicObj)
   return new_obj
 end
 
-function get_row(c::ConicObj, row::Int64)
+function get_row(c::ConicObj, row::Int)
   new_obj = ConicObj()
   for (var, coeff) in c
     new_obj[var] = coeff[row, :]
@@ -47,7 +47,7 @@ function *(v::Value, c::ConicObj)
   return new_obj
 end
 
-function promote_size(c::ConicObj, vectorized_size::Int64)
+function promote_size(c::ConicObj, vectorized_size::Int)
   new_obj = copy(c)
   for var in keys(new_obj)
     new_obj[var] = repmat(new_obj[var], vectorized_size, 1)
@@ -58,11 +58,11 @@ end
 type ConicConstr
   objs::Array{ConicObj}
   cone::Symbol
-  sizes::Array{Int64}
+  sizes::Array{Int}
 end
 
 UniqueExpMap = Dict{(Symbol, Uint64), ConicObj}
-UniqueConstrMap = Dict{(Symbol, Uint64), Int64}
+UniqueConstrMap = Dict{(Symbol, Uint64), Int}
 UniqueConstrList = Array{ConicConstr}
 
 type UniqueConicForms
